Refactor to move omronUtilities class to its own header file and renamed cpp file
