+++ omroneipApp/src/omronUtilities.cpp
